Rails: • Added source.ruby.rails scope the the snippets were it made sense. • Removed a duplicated snippet. (logging) • Changed the names of the Rails methods to keyword.function.rails.*.ruby instead of keyword.function.(actionpack|activerecord).rails.ruby
Java:
• Added scope to the most snippets, because it didn't happen till now.

NOTE: The most snippets are prefixed with a j, I didn't changed this to not break any coding habbits, but i wanted to note that thits could be changed now because wo have scopes.

Quake:
• Just some basic ready making for some upcoming stuff. :)

Ruby:
• Simplified the syntax checker.
• Removed the scope from 'Execute Line with Ruby' and 'Start up IRB' because this is stuff, I usually use from every scope and want to do this futher.
• Reworked the declaration.function pattern to catch all possible method names and also to include source.ruby in the arguments to color numbers and regex, etc corretly.  This just has 2 problems:
a) Cant find function-arguments, because i dont know how to name whats between begin and end
b) if you enter a ) in the arguments the match gets aborted to early, one could prewent this with a recursive swallow, but this doesn't perform very well on my machine so I left it to now slow down everything.

Subversion:
• Imported a basic Revert, more to come in the next days.



git-svn-id: http://svn.textmate.org/trunk/Bundles/Quake.tmbundle@763 dfb7d73b-c2ec-0310-8fea-fb051d288c6d